+++ applications/vessel-history/src/features/vessels/activity/vessels-activity.selectors.ts
@@ -2,11 +2,11 @@ import { createSelector } from '@reduxjs/toolkit'
 import { DateTime, Interval } from 'luxon'
 import {
   resolveDataviewDatasetResource,
-  resolveDataviewEventsResources,
+  resolveDataviewDatasetResources,
   selectResources,
 } from '@globalfishingwatch/dataviews-client'
 import { DatasetTypes, EventTypes, ResourceStatus } from '@globalfishingwatch/api-types'
-import { EVENTS_COLORS } from 'data/config'
+import { EVENTS_COLORS, WORKSPACE_START_DATE } from 'data/config'
 import { Filters, initialState, selectFilters } from 'features/profile/filters/filters.slice'
 import { t } from 'features/i18n/i18n'
 import { selectActiveTrackDataviews } from 'features/dataviews/dataviews.selectors'
@@ -30,7 +30,7 @@ export const selectEventsForTracks = createSelector(
     const vesselsEvents = trackDataviews.map((dataview) => {
       const { url: tracksUrl } = resolveDataviewDatasetResource(dataview, DatasetTypes.Tracks)
       // const { url: eventsUrl } = resolveDataviewDatasetResource(dataview, DatasetTypes.Events)
-      const eventsResources = resolveDataviewEventsResources(dataview)
+      const eventsResources = resolveDataviewDatasetResources(dataview, DatasetTypes.Events)
       const hasEventData =
         eventsResources?.length && eventsResources.every(({ url }) => resources[url]?.data)
       const tracksResourceResolved =
@@ -137,8 +137,8 @@ export const selectEventsWithRenderingInfo = createSelector(
             : '',
           duration.minutes && duration.minutes > 0
             ? t('event.minuteAbbreviated', '{{count}}m', {
-                count: Math.round(duration.minutes as number),
-              })
+              count: Math.round(duration.minutes as number),
+            })
             : '',
         ].join(' ')
 
@@ -206,8 +206,21 @@ const getEventRegionDescription = (event: ActivityEvent, eezs: Region[], rfmos: 
   return regionsDescription ?? ''
 }
 
+export const selectEvents = createSelector(
+  [selectEventsWithRenderingInfo],
+  (events) => events.flat().sort((a, b) => (a.start > b.start ? -1 : 1))
+)
+
+export const selectMapEvents = createSelector(
+  [selectEvents, selectFilters],
+  (events, filters) => {
+    const startDate = DateTime.fromISO(filters.start, { zone: 'utc' })
+    return events.filter(event => event.start >= startDate.toMillis())
+  }
+)
+
 export const selectFilteredEvents = createSelector(
-  [selectEventsWithRenderingInfo, selectFilters],
+  [selectMapEvents, selectFilters],
   (events, filters) => {
     // Need to parse the timerange start and end dates in UTC
     // to not exclude events in the boundaries of the range
@@ -221,7 +234,6 @@ export const selectFilteredEvents = createSelector(
     const interval = Interval.fromDateTimes(startDate, endDate)
 
     return events
-      .flat()
       .filter((event: RenderedEvent) => {
         if (
           !interval.contains(DateTime.fromMillis(event.start as number)) &&
@@ -244,7 +256,7 @@ export const selectFilteredEvents = createSelector(
 
         return true
       })
-      .sort((a, b) => (a.start > b.start ? -1 : 1))
+
   }
 )