+++ packages/timebar/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @globalfishingwatch/timebar
 
+## 2.0.8
+
+### Patch Changes
+
+- 567ee29e: fix protpypes false positives
+
 ## 2.0.7
 
 ### Patch Changes